Исправление мелких недочетов
