+++ js/modal.js
@@ -1,4 +1,4 @@
-import { resetMap } from './map.js';
+
 import { isEscapeKey } from './util.js';
 
 const errorMessage = document.querySelector('#error')
@@ -9,32 +9,10 @@ const successMessage = document.querySelector('#success')
   .content
   .querySelector('.success');
 
-const submitButton = document.querySelector('.ad-form__submit');
 
-const blockSubmitButton = () => {
-  submitButton.disabled = true;
-  submitButton.textContent = 'Публикую...';
-};
-
-const unblockSubmitButton = () => {
-  submitButton.disabled = false;
-  submitButton.textContent = 'Опубликовать';
-};
-
-const onSendError = () => {
+const showErrorMessage = () => {
   const errorPopup = errorMessage.cloneNode(true);
-  const resetBtn = errorPopup.querySelector('.error__button');
   document.body.append(errorPopup);
-  resetBtn.addEventListener('click', () => {
-    errorPopup.remove();
-  });
-  unblockSubmitButton();
-};
-
-const onSendSuccess = () => {
-  const successPopup = successMessage.cloneNode(true);
-  document.body.append(successPopup);
-
 
   const onPopupEscKeydown = (evt) => {
     if (isEscapeKey(evt)) {
@@ -47,18 +25,39 @@ const onSendSuccess = () => {
     closeUserModal();
   };
 
-  document.querySelector('.ad-form').reset();
-  resetMap();
-  unblockSubmitButton();
-
   function closeUserModal() {
-    successPopup.remove();
+    errorPopup.remove();
     document.removeEventListener('keydown', onPopupEscKeydown);
-    document.addEventListener('click', onPopupCloseClick);
+    document.removeEventListener('click', onPopupCloseClick);
   }
 
   document.addEventListener('keydown', onPopupEscKeydown);
   document.addEventListener('click', onPopupCloseClick);
 };
 
-export { onSendError, onSendSuccess, blockSubmitButton };
+const showSuccessMessage = () => {
+  const successPopup = successMessage.cloneNode(true);
+  document.body.append(successPopup);
+
+  const onPopupEscKeydown = (evt) => {
+    if (isEscapeKey(evt)) {
+      evt.preventDefault();
+      closeUserModal();
+    }
+  };
+
+  const onPopupCloseClick = () => {
+    closeUserModal();
+  };
+
+  function closeUserModal() {
+    successPopup.remove();
+    document.removeEventListener('keydown', onPopupEscKeydown);
+    document.removeEventListener('click', onPopupCloseClick);
+  }
+
+  document.addEventListener('keydown', onPopupEscKeydown);
+  document.addEventListener('click', onPopupCloseClick);
+};
+
+export { showErrorMessage, showSuccessMessage };
